test: 따로 빼놓은 지출 목록 조회 테스트도 ExpenditureControllerTest로 합침
