+++ src/test/resources/db/member-expenditure.sql
@@ -31,7 +31,30 @@ VALUES ('member1', '$2a$10$x6a5r29ElOc5nThogQozT.ZZijwjZKEODkl52ayATgVtMm5gbe4Zy
 
 INSERT INTO expenditure (member_id, category_id, amount, memo, exclude_from_total,
                          expenditure_at, created_at, updated_at)
-VALUES (1, 1, 20000, '치킨', 0, '2023-11-20 13:15:42', '2023-11-20 20:36:17', '2023-11-20 20:36:17'),
-       (1, 2, 6800, '택시비', 0, '2023-11-21 08:22:35', '2023-11-21 21:20:05', '2023-11-21 21:20:05'),
-       (2, 5, 120000, '인프런 강의 구매', 0, '2023-11-21 08:22:35', '2023-11-21 21:20:05', '2023-11-21 21:20:05'),
-       (2, 8, 49800, '23년 11월 통신비', 0, '2023-11-22 17:20:05', '2023-11-22 20:48:55', '2023-11-22 20:48:55');
+VALUES
+    -- member1
+    -- FOOD category
+    (1, 1, 10000, '음식 1', false, '2023-11-02T11:30:00', NOW(), NOW()),
+    (1, 1, 20000, '음식 2', false, '2023-11-02T12:30:00', NOW(), NOW()),
+    (1, 1, 30000, '음식 3', false, '2023-11-02T13:30:00', NOW(), NOW()),
+    (1, 1, 40000, '음식 4', false, '2023-11-02T14:30:00', NOW(), NOW()),
+    (1, 1, 50000, '음식 5', false, '2023-11-02T15:30:00', NOW(), NOW()),
+
+    -- EDUCATION category
+    (1, 5, 120000, '강의 구매 6', false, '2023-11-03T16:30:00', NOW(), NOW()),
+    (1, 5, 140000, '강의 구매 7', false, '2023-11-03T17:30:00', NOW(), NOW()),
+
+    -- member2
+    -- UNCATEGORIZED category
+    (2, 10, 80000, '기타 8', false, '2023-11-06T18:30:00', NOW(), NOW()),
+    (2, 10, 90000, '기타 9', false, '2023-11-06T19:30:00', NOW(), NOW()),
+
+    -- UNCATEGORIZED category, excludeFromTotal true
+    (2, 10, 100000, '기타 10', true, '2023-11-07T20:30:00', NOW(), NOW()),
+    (2, 10, 110000, '기타 11', true, '2023-11-07T21:30:00', NOW(), NOW()),
+
+    -- SHOPPING category
+    (2, 3, 12000, '쇼핑 12', false, '2023-11-09T10:12:00', NOW(), NOW()),
+    (2, 3, 13000, '쇼핑 13', false, '2023-11-09T10:13:00', NOW(), NOW()),
+    (2, 3, 14000, '쇼핑 14', false, '2023-11-09T10:14:00', NOW(), NOW()),
+    (2, 3, 15000, '쇼핑 15', false, '2023-11-09T10:15:00', NOW(), NOW());
